requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
sqlalchemy==2.0.0
passlib[bcrypt]==1.7.4
python-dotenv==0.21.0
pydantic==1.10.9
python-jose==3.3.0

